# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Error {
  error: string;
}

export type CrmContactStatus = typeof CrmContactStatus[keyof typeof CrmContactStatus];


export const CrmContactStatus = {
  lead: 'lead',
  active: 'active',
  partner: 'partner',
  dormant: 'dormant',
} as const;

export interface CrmContact {
  id: number;
  name: string;
  role: string;
  area: string;
  affiliation: string;
  status: CrmContactStatus;
  engagementScore: number;
  /** @nullable */
  lastContactedAt: string | null;
  nextAction: string;
  notes: string;
  createdAt: string;
  updatedAt: string;
}

export type CrmContactInputStatus = typeof CrmContactInputStatus[keyof typeof CrmContactInputStatus];


export const CrmContactInputStatus = {
  lead: 'lead',
  active: 'active',
  partner: 'partner',
  dormant: 'dormant',
} as const;

export interface CrmContactInput {
  /** @minLength 1 */
  name: string;
  /** @minLength 1 */
  role: string;
  /** @minLength 1 */
  area: string;
  /** @minLength 1 */
  affiliation: string;
  status: CrmContactInputStatus;
  /**
     * @minimum 0
     * @maximum 100
     */
  engagementScore: number;
  /** @nullable */
  lastContactedAt?: string | null;
  nextAction: string;
  notes: string;
}

export type CrmContactUpdateStatus = typeof CrmContactUpdateStatus[keyof typeof CrmContactUpdateStatus];


export const CrmContactUpdateStatus = {
  lead: 'lead',
  active: 'active',
  partner: 'partner',
  dormant: 'dormant',
} as const;

export interface CrmContactUpdate {
  status?: CrmContactUpdateStatus;
  /**
     * @minimum 0
     * @maximum 100
     */
  engagementScore?: number;
  /** @nullable */
  lastContactedAt?: string | null;
  nextAction?: string;
  notes?: string;
}

export type CrmTaskStatus = typeof CrmTaskStatus[keyof typeof CrmTaskStatus];


export const CrmTaskStatus = {
  todo: 'todo',
  'in-progress': 'in-progress',
  done: 'done',
} as const;

export type CrmTaskPriority = typeof CrmTaskPriority[keyof typeof CrmTaskPriority];


export const CrmTaskPriority = {
  low: 'low',
  medium: 'medium',
  high: 'high',
} as const;

export interface CrmTask {
  id: number;
  title: string;
  status: CrmTaskStatus;
  priority: CrmTaskPriority;
  /** @nullable */
  dueDate: string | null;
  owner: string;
  /** @nullable */
  contactId: number | null;
  createdAt: string;
  updatedAt: string;
}

export type CrmTaskInputStatus = typeof CrmTaskInputStatus[keyof typeof CrmTaskInputStatus];


export const CrmTaskInputStatus = {
  todo: 'todo',
  'in-progress': 'in-progress',
  done: 'done',
} as const;

export type CrmTaskInputPriority = typeof CrmTaskInputPriority[keyof typeof CrmTaskInputPriority];


export const CrmTaskInputPriority = {
  low: 'low',
  medium: 'medium',
  high: 'high',
} as const;

export interface CrmTaskInput {
  /** @minLength 1 */
  title: string;
  status: CrmTaskInputStatus;
  priority: CrmTaskInputPriority;
  /** @nullable */
  dueDate?: string | null;
  /** @minLength 1 */
  owner: string;
  /** @nullable */
  contactId?: number | null;
}

export type CrmTaskUpdateStatus = typeof CrmTaskUpdateStatus[keyof typeof CrmTaskUpdateStatus];


export const CrmTaskUpdateStatus = {
  todo: 'todo',
  'in-progress': 'in-progress',
  done: 'done',
} as const;

export type CrmTaskUpdatePriority = typeof CrmTaskUpdatePriority[keyof typeof CrmTaskUpdatePriority];


export const CrmTaskUpdatePriority = {
  low: 'low',
  medium: 'medium',
  high: 'high',
} as const;

export interface CrmTaskUpdate {
  status?: CrmTaskUpdateStatus;
  priority?: CrmTaskUpdatePriority;
  /** @nullable */
  dueDate?: string | null;
  owner?: string;
  /** @nullable */
  contactId?: number | null;
}

export type CommunitySignalStatus = typeof CommunitySignalStatus[keyof typeof CommunitySignalStatus];


export const CommunitySignalStatus = {
  new: 'new',
  triaged: 'triaged',
  'in-progress': 'in-progress',
  resolved: 'resolved',
} as const;

export interface CommunitySignal {
  id: number;
  area: string;
  summary: string;
  status: CommunitySignalStatus;
  source: string;
  communityVisible: boolean;
  createdAt: string;
  updatedAt: string;
}

export interface CommunitySignalInput {
  /** @minLength 1 */
  area: string;
  /**
     * @minLength 1
     * @maxLength 1000
     */
  summary: string;
}

export type CommunitySignalUpdateStatus = typeof CommunitySignalUpdateStatus[keyof typeof CommunitySignalUpdateStatus];


export const CommunitySignalUpdateStatus = {
  new: 'new',
  triaged: 'triaged',
  'in-progress': 'in-progress',
  resolved: 'resolved',
} as const;

export interface CommunitySignalUpdate {
  status?: CommunitySignalUpdateStatus;
  communityVisible?: boolean;
}

export interface CrmSummary {
  totalContacts: number;
  activeContacts: number;
  openTasks: number;
  highPriorityTasks: number;
  newSignals: number;
  averageEngagement: number;
}

export interface CrmOverview {
  summary: CrmSummary;
  contacts: CrmContact[];
  tasks: CrmTask[];
  signals: CommunitySignal[];
}

/**
 * Resource not found
 */
export type NotFoundResponse = Error;