# lib/api-zod/src/generated/types/communitySignal.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { CommunitySignalStatus } from './communitySignalStatus';

export interface CommunitySignal {
  id: number;
  area: string;
  summary: string;
  status: CommunitySignalStatus;
  source: string;
  communityVisible: boolean;
  createdAt: Date;
  updatedAt: Date;
}