# lib/api-zod/src/generated/types/communitySignalInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface CommunitySignalInput {
  /** @minLength 1 */
  area: string;
  /**
     * @minLength 1
     * @maxLength 1000
     */
  summary: string;
}