# lib/api-zod/src/generated/types/communitySignalUpdate.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { CommunitySignalUpdateStatus } from './communitySignalUpdateStatus';

export interface CommunitySignalUpdate {
  status?: CommunitySignalUpdateStatus;
  communityVisible?: boolean;
}