# lib/api-zod/src/generated/types/crmContact.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { CrmContactStatus } from './crmContactStatus';

export interface CrmContact {
  id: number;
  name: string;
  role: string;
  area: string;
  affiliation: string;
  status: CrmContactStatus;
  engagementScore: number;
  /** @nullable */
  lastContactedAt: Date | null;
  nextAction: string;
  notes: string;
  createdAt: Date;
  updatedAt: Date;
}