# lib/api-zod/src/generated/types/crmContactInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { CrmContactInputStatus } from './crmContactInputStatus';

export interface CrmContactInput {
  /** @minLength 1 */
  name: string;
  /** @minLength 1 */
  role: string;
  /** @minLength 1 */
  area: string;
  /** @minLength 1 */
  affiliation: string;
  status: CrmContactInputStatus;
  /**
     * @minimum 0
     * @maximum 100
     */
  engagementScore: number;
  /** @nullable */
  lastContactedAt?: Date | null;
  nextAction: string;
  notes: string;
}