# lib/api-zod/src/generated/types/crmContactUpdate.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { CrmContactUpdateStatus } from './crmContactUpdateStatus';

export interface CrmContactUpdate {
  status?: CrmContactUpdateStatus;
  /**
     * @minimum 0
     * @maximum 100
     */
  engagementScore?: number;
  /** @nullable */
  lastContactedAt?: Date | null;
  nextAction?: string;
  notes?: string;
}