# lib/api-zod/src/generated/types/crmOverview.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { CommunitySignal } from './communitySignal';
import type { CrmContact } from './crmContact';
import type { CrmSummary } from './crmSummary';
import type { CrmTask } from './crmTask';

export interface CrmOverview {
  summary: CrmSummary;
  contacts: CrmContact[];
  tasks: CrmTask[];
  signals: CommunitySignal[];
}