# lib/api-zod/src/generated/types/crmSummary.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface CrmSummary {
  totalContacts: number;
  activeContacts: number;
  openTasks: number;
  highPriorityTasks: number;
  newSignals: number;
  averageEngagement: number;
}