# lib/api-zod/src/generated/types/crmTask.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { CrmTaskPriority } from './crmTaskPriority';
import type { CrmTaskStatus } from './crmTaskStatus';

export interface CrmTask {
  id: number;
  title: string;
  status: CrmTaskStatus;
  priority: CrmTaskPriority;
  /** @nullable */
  dueDate: Date | null;
  owner: string;
  /** @nullable */
  contactId: number | null;
  createdAt: Date;
  updatedAt: Date;
}