# lib/api-zod/src/generated/types/crmTaskInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { CrmTaskInputPriority } from './crmTaskInputPriority';
import type { CrmTaskInputStatus } from './crmTaskInputStatus';

export interface CrmTaskInput {
  /** @minLength 1 */
  title: string;
  status: CrmTaskInputStatus;
  priority: CrmTaskInputPriority;
  /** @nullable */
  dueDate?: Date | null;
  /** @minLength 1 */
  owner: string;
  /** @nullable */
  contactId?: number | null;
}