# lib/api-zod/src/generated/types/crmTaskUpdate.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { CrmTaskUpdatePriority } from './crmTaskUpdatePriority';
import type { CrmTaskUpdateStatus } from './crmTaskUpdateStatus';

export interface CrmTaskUpdate {
  status?: CrmTaskUpdateStatus;
  priority?: CrmTaskUpdatePriority;
  /** @nullable */
  dueDate?: Date | null;
  owner?: string;
  /** @nullable */
  contactId?: number | null;
}